move functions for interact with btc_connector
