+++ bridge-sdk/bridge-clients/near-bridge-client/src/near_bridge_client.rs
@@ -4,16 +4,16 @@ use bridge_connector_common::result::{BridgeSdkError, Result};
 use derive_builder::Builder;
 use near_contract_standards::storage_management::StorageBalance;
 use near_crypto::{SecretKey, Signer};
-use near_primitives::types::Gas;
 use near_primitives::{hash::CryptoHash, types::AccountId, views::TxExecutionStatus};
 use near_rpc_client::{ChangeRequest, ViewRequest};
-use near_sdk::json_types::U128;
 use near_token::NearToken;
 use omni_types::{
     locker_args::{BindTokenArgs, ClaimFeeArgs, DeployTokenArgs, FinTransferArgs},
     ChainKind, Fee, OmniAddress, TransferId, TransferMessage,
 };
 use serde_json::json;
+
+pub mod btc_connector;
 
 const STORAGE_DEPOSIT_GAS: u64 = 10_000_000_000_000;
 
@@ -32,8 +32,6 @@ const FT_TRANSFER_DEPOSIT: u128 = 1;
 
 const FIN_TRANSFER_GAS: u64 = 300_000_000_000_000;
 const FIN_TRANSFER_DEPOSIT: u128 = 600_000_000_000_000_000_000;
-
-const FIN_BTC_TRANSFER_GAS: u64 = 300_000_000_000_000;
 
 const CLAIM_FEE_GAS: u64 = 300_000_000_000_000;
 const CLAIM_FEE_DEPOSIT: u128 = 1;
@@ -62,34 +60,6 @@ impl Default for TransactionOptions {
 #[derive(serde::Deserialize)]
 struct StorageBalanceBounds {
     min: NearToken,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub struct PostAction {
-    pub receiver_id: AccountId,
-    pub amount: U128,
-    pub memo: Option<String>,
-    pub msg: String,
-    pub gas: Option<Gas>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub struct DepositMsg {
-    pub recipient_id: AccountId,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub post_actions: Option<Vec<PostAction>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extra_msg: Option<String>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub struct FinBtcTransferArgs {
-    pub deposit_msg: DepositMsg,
-    pub tx_bytes: Vec<u8>,
-    pub vout: usize,
-    pub tx_block_blockhash: String,
-    pub tx_index: u64,
-    pub merkle_proof: Vec<String>,
 }
 
 #[derive(serde::Serialize)]
@@ -664,37 +634,6 @@ impl NearBridgeClient {
         tracing::info!(
             tx_hash = tx_hash.to_string(),
             "Sent finalize transfer transaction"
-        );
-        Ok(tx_hash)
-    }
-
-    pub async fn fin_btc_transfer(
-        &self,
-        args: FinBtcTransferArgs,
-        transaction_options: TransactionOptions,
-        wait_final_outcome_timeout_sec: Option<u64>,
-    ) -> Result<CryptoHash> {
-        let endpoint = self.endpoint()?;
-        let btc_connector = self.btc_connector()?;
-        let tx_hash = near_rpc_client::change_and_wait(
-            endpoint,
-            ChangeRequest {
-                signer: self.signer()?,
-                nonce: transaction_options.nonce,
-                receiver_id: btc_connector,
-                method_name: "verify_deposit".to_string(),
-                args: serde_json::json!(args).to_string().into_bytes(),
-                gas: FIN_BTC_TRANSFER_GAS,
-                deposit: 0,
-            },
-            transaction_options.wait_until,
-            wait_final_outcome_timeout_sec,
-        )
-        .await?;
-
-        tracing::info!(
-            tx_hash = tx_hash.to_string(),
-            "Sent BTC finalize transfer transaction"
         );
         Ok(tx_hash)
     }
